Show tooltip only if the quality warning has not been recently shown
In some cases the warning could be repeteadly shown and hidden again. In
those cases showing the warning is fine, but showing the tooltip could
be annoying. Therefore now the tooltip is automatically shown only if
the warning has not been recently shown. In any case, the tooltip can
still be shown when hovering on the warning.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>